Port Actor and Player to 328.8; fix struct field type mismatches
- ActorComponentOffsets: regenerated to 328.8 (action/animation/skills/deployed
  object arrays at real offsets); field names preserved so Actor.cs is unchanged
  apart from the destination conversion below.
- ActionWrapperOffsets.Destination is a Vector2i (grid) in 328.8, not a float
  Vector2; fixed the type and Actor.ActionWrapper.Destination now converts.
- Player: updated the dump-confirmed offsets (name 0x168, xp 0x18C, attributes
  0x19C/0x1A0/0x1A4, level 0x1AC).
- Type fixes flagged by a repo-vs-dump audit: ServerStashTabOffsets Member/Officer
  flags are bytes (were uint and overlapped adjacent fields); CursorOffsets.Action
  is a byte.

GameOffsets + Core validate for syntax with Roslyn (0 errors).

+++ GameOffsets/CursorOffsets.cs
@@ -17,7 +17,7 @@ public struct CursorOffsets
     [FieldOffset(0x0)] public int vTable;
 
     /// <summary>Current cursor action identifier.</summary>
-    [FieldOffset(0x578)] public int Action;
+    [FieldOffset(0x578)] public byte Action;
 
     /// <summary>Number of clicks registered.</summary>
     [FieldOffset(0x2CC)] public int Clicks;
